backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pymongo==4.6.0
sentence-transformers==2.2.2
langchain==0.1.0
langchain-community==0.0.10
huggingface-hub==0.19.4
transformers==4.36.0
torch>=2.0.0
pypdf==3.17.1
python-multipart==0.0.6
python-dotenv==1.0.0
einops==0.8.1
